Add CUDA + PyTorch
Assumes CUDA 12.0 is installed. Might work with other versions.

Note that installing pytorch is by far the longest
part of this process. By using an image with pytorch pre-installed,
this could build much faster. However, copying torch from an image
would be pretty hard, so it might be worth copying/installing ROS
on top of a pre-built torch container.