Add CUDA + PyTorch (#31)
* Add CUDA + PyTorch

Assumes CUDA 12.0 is installed. Might work with other versions.

Note that installing pytorch is by far the longest
part of this process. By using an image with pytorch pre-installed,
this could build much faster. However, copying torch from an image
would be pretty hard, so it might be worth copying/installing ROS
on top of a pre-built torch container.

* Build with/without GPU via arguments

Allows for disabling (--no-gpu) GPU, or forcing
(--force-gpu) GPU on platforms without a GPU.

Development environment defaults to use no GPU if
nvidia container toolkit is not present. Production
will throw a warning and fail if this is the case
and --no-gpu is not explicitly specified.

* Lock down all CUDA requirement versions